requirements.txt
python-telegram-bot==13.11
environs==9.5.0